requirements.txt
lxml==5.3.0
numpy==2.1.3
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
accelerate==1.3.0
diffusers==0.32.2
huggingface-hub==0.27.1
protobuf==5.29.3
sentencepiece==0.2.0
torch==2.5.1
transformers==4.48.1
rembg==2.0.61
onnxruntime
selenium
cairosvg
openai